A small example how we are doing it in our current project.
I have renamed PublishingServiceTest to PublishingServiceShould to express the behavior more.

Now you can read it more like PublishingServiceShould -> PublishAnArticle.

I'm using autofac to resolve the service and i don't have to worry about the wiring of the repositories and the services.

In my current project we are using an autofac module that we can reuse in the test. So if we add a new service to our handler we don't have to worry about the wiring.

In the 'PublishingServiceShould' i'm using a TestContextBuilder to setup my TestContext. With that approach we can create a different context per Test in the class.

For TeaserServiceTest I'm inheriting from an BaseClass as an example. Not a big fan of a baseclass because the testcontext reflects the entire testclass and is more difficult to set it per test.

Looking forward for your feedback.